% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_weeks.R
\name{aggregate_weeks}
\alias{aggregate_weeks}
\title{Aggregate daily state-level data into week-level intervals}
\usage{
aggregate_weeks(dat)
}
\arguments{
\item{dat}{A data frame of downloaded and cleaned state-level data from covidtracking.com, with absent negatives collapsed.}
}
\value{
A data frame grouped by state. Contains weekly aggregated counts of positive tests, negative tests, death, and hospitalizations.
Time is measured in epidemiological weeks since 2020-03-01.
}
\description{
Aggregate daily state-level data into week-level intervals
}
\examples{
dat <- pull_dat()
dat2 <- clean_dat(dat)
dat3 <- collapse_absent_negs(dat2)
dat4 <- aggregate_weeks(dat3)
}
