% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dat.R
\name{clean_dat}
\alias{clean_dat}
\title{Take raw daily state data, clean, and retain incidence}
\usage{
clean_dat(dat)
}
\arguments{
\item{dat}{A data frame of downloaded state-level data from covidtracking.com}
}
\value{
A data frame grouped by state. Contains daily counts of positive tests, negative tests, death, and hospitalizations.
}
\description{
Take raw daily state data, clean, and retain incidence
}
\examples{
dat <- pull_dat()
dat2 <- clean_dat(dat)
}
