% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_absent_negs.R
\name{collapse_absent_negs}
\alias{collapse_absent_negs}
\title{Sum positive cases over dates without negative data}
\usage{
collapse_absent_negs(dat)
}
\arguments{
\item{dat}{A data frame of downloaded and cleaned state-level data from covidtracking.com}
}
\value{
A data frame grouped by state. Contains daily counts of positive tests, negative tests, death, and hospitalizations. For dates
For those dates without negative incidence data in a given state, this function sums up the positive cases over those dates
and adds this sum to the number of positive cases in the next date which does contain negative incidence.
}
\description{
Sum positive cases over dates without negative data
}
\examples{
dat <- pull_dat()
dat2 <- clean_dat(dat)
dat3 <- collapse_absent_negs(dat2)
}
