% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t0.R
\name{get_t0}
\alias{get_t0}
\title{Add row to weekly data to denote the day before first observation}
\usage{
get_t0(dat_daily, dat_weekly)
}
\arguments{
\item{dat_daily}{A data frame of downloaded and cleaned state-level data from covidtracking.com}

\item{dat_weekly}{An aggregated weekly-level dataset output by other functions in this package.}
}
\value{
A data frame grouped by state. Contains weekly aggregated counts of positive tests, negative tests, death, and hospitalizations.
Time is measured in epidemiological weeks since 2020-03-01. The first row in each state corresponds to the day before the first
observation in that state.
}
\description{
Add row to weekly data to denote the day before first observation
}
\examples{
dat <- pull_dat()
dat2 <- clean_dat(dat)
dat3 <- collapse_absent_negs(dat2)
dat4 <- aggregate_weeks(dat3)
dat5 <- get_t0(dat, dat4)
}
