% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dat.R
\name{pull_dat}
\alias{pull_dat}
\title{Download daily state-level data from covidtracking.com}
\usage{
pull_dat()
}
\value{
A data frame containing recent daily state-level data on various covariates
}
\description{
Download daily state-level data from covidtracking.com
}
\examples{
dat <- pull_dat()
}
