% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_or_allNA.R
\name{sum_or_allNA}
\alias{sum_or_allNA}
\title{Sum a vector or return NA if all entries are NA}
\usage{
sum_or_allNA(x)
}
\arguments{
\item{x}{A vector containing numerics and NAs}
}
\value{
Either the sum of non-NA elements of the input vector, or NA if all elements of the vector are NA
}
\description{
Sum a vector or return NA if all entries are NA
}
\examples{
x1 <- c(1,2,3,NA)
sum_or_allNA(x1) # returns 6

x2 <- c(NA, NA, NA, NA)
sum_or_allNA(x2) # returns NA
}
