% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_state_dat.R
\name{get_weekly_state_dat}
\alias{get_weekly_state_dat}
\title{Get aggregated weekly state-level data from covidtracking.com}
\usage{
get_weekly_state_dat()
}
\value{
A data frame grouped by state. Contains weekly aggregated counts of positive tests, negative tests, death, and hospitalizations.
For those dates without negative incidence data in a given state, the sum of the positive cases over those dates
is added to the number of positive cases in the next date which does contain negative incidence. The t0 column for a given
state represents the day before covidtracking.com had data for that state.
}
\description{
Get aggregated weekly state-level data from covidtracking.com
}
\examples{
dat <- get_weekly_state_dat()
}
